UI/requirements.txt
Werkzeug
gunicorn
Flask
google-cloud-storage

